% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_function.R
\name{cbmsUtils}
\alias{cbmsUtils}
\title{cbmsUtils Function}
\usage{
cbmsUtils(main)
}
\arguments{
\item{main}{CBMS CORE HPQ VN05201701 main.csv from your .can file. Can be a path don't forget to quote it or store it in variable.}
}
\description{
This function allows you to generate/segregate: duplicate files, unfiltered files and distinct files.
}
\examples{
cbmsUtils('main.csv')

cbmsUtils('"C:/Users/name/main.csv"')
}
\keyword{cbms}
