% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusTab_function.R
\name{fusTab}
\alias{fusTab}
\title{fusTab Function}
\usage{
fusTab(main)
}
\arguments{
\item{main}{CBMS CORE HPQ VN05201701 main.csv from your .can file. Can be a path don't forget to quote it or store it in variable.}
}
\description{
This function you to create a geopoint and lite version of main.csv for gps tracking
}
\examples{
fusTab('main.csv')

fusTab('"C:/Users/name/main.csv"')
}
\keyword{cbms}
