% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popSum_function.R
\name{popSum}
\alias{popSum}
\title{popSum Function}
\usage{
popSum(main)
}
\arguments{
\item{main}{CBMS CORE HPQ VN05201701 main.csv from your .can file. Can be a path don't forget to quote it or store it in variable.}
}
\description{
This function allows you to generate a summary of Population and Household in cbms.
}
\examples{
popSum('main.csv')

popSum('"C:/Users/name/main.csv"')
}
\keyword{cbms}
