% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCBMS.R
\name{readCBMS}
\alias{readCBMS}
\title{readCBMS Function}
\usage{
readCBMS(x)
}
\arguments{
\item{fread}{for faster data reading and larger data}

\item{read.csv}{for small data}
}
\description{
This function reads all file in a directory.
}
\examples{
readCBMS('fread')

readCBMS('read.csv')
}
\keyword{cbms}
